gateway-api/requirements.txt
# Base, core, framework
starlette # 0.37.2
fastapi==0.111.0
uvicorn # 0.29.0
pydantic==2.6.4
load_dotenv # 0.1.0
pytz # 2024.1
python-multipart # 0.0.9
slowapi # 0.1.9
pyjwt[crypto] # 2.8.0
httpx # 0.27.0

# Logging
sentry-sdk # 2.0.1

# DB
prisma

# CLI
dotenv-cli

# TEST
pytest # 8.2.0
pytest-asyncio # 0.23.6
pytest-order # 1.2.1
